+++ client/jetbrains/src/main/java/com/sourcegraph/cody/autocomplete/CodyAutocompleteManager.java
@@ -9,7 +9,6 @@ import com.intellij.openapi.fileEditor.FileDocumentManager;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Disposer;
 import com.intellij.openapi.util.TextRange;
-import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.codeStyle.CommonCodeStyleSettings;
 import com.intellij.util.concurrency.annotations.RequiresEdt;
 import com.sourcegraph.cody.agent.CodyAgent;
@@ -122,7 +121,7 @@ public class CodyAutocompleteManager {
    * @param offset The character offset in the editor to trigger auto-complete at.
    */
   public void triggerAutocomplete(
-      @NotNull Editor editor, int offset, InlineCompletionTriggerKind triggerKind) {
+      @NotNull Editor editor, int offset, @NotNull InlineCompletionTriggerKind triggerKind) {
     boolean isTriggeredManually = triggerKind.equals(InlineCompletionTriggerKind.INVOKE);
     String currentCommand = CommandProcessor.getInstance().getCurrentCommandName();
     if (!ConfigUtil.isCodyEnabled()) return;
@@ -131,7 +130,7 @@ public class CodyAutocompleteManager {
       return;
     } else if (!isTriggeredManually
         && !CodyEditorUtil.isImplicitAutocompleteEnabledForEditor(editor)) return;
-    else if (CodyEditorUtil.isCommandExcluded(currentCommand)) {
+    else if (CodyEditorUtil.isCommandExcluded(currentCommand) && !isTriggeredManually) {
       return;
     }
     final Project project = editor.getProject();
@@ -168,8 +167,8 @@ public class CodyAutocompleteManager {
       @NotNull Editor editor,
       int offset,
       @NotNull TextDocument textDocument,
-      InlineCompletionTriggerKind triggerKind,
-      CancellationToken cancellationToken) {
+      @NotNull InlineCompletionTriggerKind triggerKind,
+      @NotNull CancellationToken cancellationToken) {
     CodyAgentServer server = CodyAgent.getServer(project);
     boolean isAgentAutocomplete = server != null;
     if (!isAgentAutocomplete) {
@@ -219,9 +218,9 @@ public class CodyAutocompleteManager {
       @NotNull Project project,
       @NotNull Editor editor,
       int offset,
-      InlineCompletionTriggerKind triggerKind,
-      InlineAutocompleteList result,
-      CancellationToken cancellationToken) {
+      @NotNull InlineCompletionTriggerKind triggerKind,
+      @NotNull InlineAutocompleteList result,
+      @NotNull CancellationToken cancellationToken) {
     if (currentAutocompleteTelemetry != null) {
       currentAutocompleteTelemetry.markCompletionEvent(result.completionEvent);
     }
@@ -281,14 +280,13 @@ public class CodyAutocompleteManager {
   private void displayAgentAutocomplete(
       @NotNull Editor editor,
       int offset,
-      InlineAutocompleteItem item,
-      InlayModel inlayModel,
-      InlineCompletionTriggerKind triggerKind) {
+      @NotNull InlineAutocompleteItem item,
+      @NotNull InlayModel inlayModel,
+      @NotNull InlineCompletionTriggerKind triggerKind) {
     TextRange range = CodyEditorUtil.getTextRange(editor.getDocument(), item.range);
     String originalText = editor.getDocument().getText(range);
     String insertTextFirstLine = item.insertText.lines().findFirst().orElse("");
-    String multilineInsertText =
-        item.insertText.lines().skip(1).collect(Collectors.joining("\n"));
+    String multilineInsertText = item.insertText.lines().skip(1).collect(Collectors.joining("\n"));
 
     // Run Myer's diff between the existing text in the document and the first line of the
     // `insertText` that is returned from the agent.