+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/auth/AccountDetails.kt
@@ -1,0 +1,8 @@
+package com.sourcegraph.cody.auth
+
+import com.intellij.openapi.util.NlsSafe
+
+interface AccountDetails {
+  @get:NlsSafe
+  val name: String
+}