+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/auth/AccountsRepository.kt
@@ -1,0 +1,8 @@
+package com.sourcegraph.cody.auth
+
+/**
+ * In most cases should be an instance of [com.intellij.openapi.components.PersistentStateComponent]
+ */
+interface AccountsRepository<A: Account> {
+  var accounts: Set<A>
+}