+++ internal/authz/providers/bitbucketcloud/provider_test.go
@@ -11,13 +11,15 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"golang.org/x/oauth2"
+
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbmocks"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketcloud"
+	"github.com/sourcegraph/sourcegraph/internal/ratelimit"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 	"github.com/sourcegraph/sourcegraph/schema"
-	"golang.org/x/oauth2"
 )
 
 type mockDoer struct {
@@ -74,6 +76,8 @@ func createTestServer() *httptest.Server {
 }
 
 func TestProvider_FetchUserPerms(t *testing.T) {
+	ratelimit.SetupForTest(t)
+
 	db := dbmocks.NewMockDB()
 	t.Run("nil account", func(t *testing.T) {
 		p := NewProvider(db,
@@ -183,6 +187,8 @@ func TestProvider_FetchUserPerms(t *testing.T) {
 }
 
 func TestProvider_FetchRepoPerms(t *testing.T) {
+	ratelimit.SetupForTest(t)
+
 	server := createTestServer()
 	defer server.Close()
 	db := dbmocks.NewMockDB()