+++ internal/search/exhaustive/service/search.go
@@ -1,14 +1,15 @@
 package service
 
 import (
+	"bytes"
 	"context"
 	"encoding/csv"
 	"fmt"
-	"io"
 	"strconv"
 	"strings"
 
 	"github.com/sourcegraph/sourcegraph/internal/search/exhaustive/types"
+	"github.com/sourcegraph/sourcegraph/internal/uploadstore"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -83,48 +84,139 @@ type CSVWriter interface {
 	WriteRow(...string) error
 }
 
-func NewCSVWriterFake(w io.Writer) CSVWriter {
-	return csvWriterFake{
-		w: csv.NewWriter(w),
+// NewBlobstoreCSVWriter creates a new BlobstoreCSVWriter which writes a CSV to
+// the store. BlobstoreCSVWriter takes care of chunking the CSV into blobs of
+// 100MiB, each with the same header row. Blobs are named {prefix}-{shard}
+// except for the first blob, which is named {prefix}.
+//
+// Data is buffered in memory until the blob reaches the maximum allowed size,
+// at which point the blob is uploaded to the store.
+//
+// The caller is expected to call Close() once and only once after the last call
+// to WriteRow.
+func NewBlobstoreCSVWriter(ctx context.Context, store uploadstore.Store, prefix string) *BlobstoreCSVWriter {
+
+	c := &BlobstoreCSVWriter{
+		maxBlobSizeBytes: 100 * 1024 * 1024,
+		ctx:              ctx,
+		prefix:           prefix,
+		store:            store,
+		// Start with "1" because we increment it before creating a new file. The second
+		// shard will be called {prefix}-2.
+		shard: 1,
 	}
+
+	c.startNewFile(ctx, prefix)
+
+	return c
 }
 
-type csvWriterFake struct {
+type BlobstoreCSVWriter struct {
+	// ctx is the context we use for uploading blobs.
+	ctx context.Context
+
+	maxBlobSizeBytes int64
+
+	prefix string
+
+	w *csv.Writer
+
+	// local buffer for the current blob.
+	buf bytes.Buffer
+
+	store uploadstore.Store
+
+	// header keeps track of the header we write as the first row of a new file.
 	header []string
-	w      *csv.Writer
+
+	// close takes care of flushing w and closing the upload.
+	close func() error
+
+	// n is the total number of bytes we have buffered so far.
+	n int64
+
+	// shard is incremented before we create a new shard.
+	shard int
 }
 
-func (c csvWriterFake) writeAndFlush(s []string) error {
+func (c *BlobstoreCSVWriter) WriteHeader(s ...string) error {
+	if c.header == nil {
+		c.header = s
+	}
+
+	// Check that c.header matches s.
+	if len(c.header) != len(s) {
+		return errors.Errorf("header mismatch: %v != %v", c.header, s)
+	}
+	for i := range c.header {
+		if c.header[i] != s[i] {
+			return errors.Errorf("header mismatch: %v != %v", c.header, s)
+		}
+	}
+
+	return c.write(s)
+}
+
+func (c *BlobstoreCSVWriter) WriteRow(s ...string) error {
+	// Create new file if we've exceeded the max blob size.
+	if c.n >= c.maxBlobSizeBytes {
+		// Close the current upload.
+		err := c.Close()
+		if err != nil {
+			return errors.Wrapf(err, "error closing upload")
+		}
+
+		c.shard++
+		c.startNewFile(c.ctx, fmt.Sprintf("%s-%d", c.prefix, c.shard))
+		err = c.WriteHeader(c.header...)
+		if err != nil {
+			return errors.Wrapf(err, "error writing header for new file")
+		}
+	}
+
+	return c.write(s)
+}
+
+// startNewFile creates a new blob and sets up the CSV writer to write to it.
+//
+// The caller is expected to call c.Close() before calling startNewFile if a
+// previous file was open.
+func (c *BlobstoreCSVWriter) startNewFile(ctx context.Context, key string) {
+	c.buf = bytes.Buffer{}
+	csvWriter := csv.NewWriter(&c.buf)
+
+	closeFn := func() error {
+		csvWriter.Flush()
+		_, err := c.store.Upload(ctx, key, &c.buf)
+		return err
+	}
+
+	c.w = csvWriter
+	c.close = closeFn
+	c.n = 0
+}
+
+// write wraps Write to keep track of the number of bytes written. This is
+// mainly for test purposes: The CSV writer is buffered (default 4096 bytes),
+// and we don't have access to the number of bytes in the buffer. In production,
+// we could just wrap the io.Pipe writer with a counter, ignore the buffer, and
+// accept that size of the blobs is off by a few kilobytes.
+func (c *BlobstoreCSVWriter) write(s []string) error {
 	err := c.w.Write(s)
 	if err != nil {
 		return err
 	}
-	c.w.Flush()
-	return c.w.Error()
-}
 
-func (c csvWriterFake) WriteHeader(s ...string) error {
-	// assert the header hasn't changed since the first call
-	if c.header == nil {
-		// first call to WriteHeader
-		c.header = s
-	} else {
-		if len(c.header) != len(s) {
-			return errors.Errorf("header mismatch: %v != %v", c.header, s)
-		}
-		for i := range c.header {
-			if c.header[i] != s[i] {
-				return errors.Errorf("header mismatch: %v != %v", c.header, s)
-			}
-		}
+	for _, field := range s {
+		c.n += int64(len(field))
 	}
+	c.n += int64(len(s)) // len(s)-1 for the commas, +1 for the newline
 
-	return c.writeAndFlush(s)
+	return nil
 }
 
-func (c csvWriterFake) WriteRow(s ...string) error {
-	return c.writeAndFlush(s)
-
+func (c *BlobstoreCSVWriter) Close() error {
+	return c.close()
 }
 
 // NewSearcherFake is a convenient working implementation of SearchQuery which